[CHUNK LOAD] Fix coquille merge

+++ srcs/SubChunk.cpp
@@ -142,7 +142,7 @@ void SubChunk::loadBiome()
 	{
 		for (int z = 0; z < CHUNK_SIZE ; z++)
 		{
-			size_t ground = (*_heightMap)[z * CHUNK_SIZE + x];
+			double ground = (*_heightMap)[z * CHUNK_SIZE + x];
 			if (ground <= OCEAN_HEIGHT)
 				loadOcean(x, z, ground);
 			else if (ground >= MOUNT_HEIGHT + (noisegen.noise(x + _position.x * CHUNK_SIZE, z + _position.z * CHUNK_SIZE) * 15)) {
